fixs for type, and add completely jobs

+++ src/Migrations/2018_11_07_024723_create_character_industry_jobs_table.php
@@ -35,7 +35,7 @@ class CreateCharacterIndustryJobsTable extends Migration
                 $table->integer('product_type_id')->nullable();
                 $table->integer('runs');
                 $table->dateTimeTz('start_date');
-                $table->integer('station_id');
+                $table->bigInteger('station_id');
                 $table->string('status');
                 $table->integer('successful_runs')->nullable();
                 $table->timestamps();
